refactor: split the code into two files-main and core
